backend/requirements.txt
flask
flask-cors
flask-login
sqlalchemy
pymysql
pandas
openpyxl
cryptography
bcrypt
pyjwt
